ruv_fann::activation

Enum ActivationFunction

pub enum ActivationFunction {

    Linear,
    Threshold,
    ThresholdSymmetric,
    Sigmoid,
    SigmoidSymmetric,
    Tanh,
    Gaussian,
    GaussianSymmetric,
    Elliot,
    ElliotSymmetric,
    LinearPiece,
    LinearPieceSymmetric,
    ReLU,
    ReLULeaky,
    Sin,
    Cos,
    SinSymmetric,
    CosSymmetric,

}

Activation functions available for neurons

These functions are based on the FANN libraryโs activation functions and include both common neural network activation functions and some specialized variants.

Variants

Linear

Linear activation function: f(x) = x * steepness

Threshold

Threshold activation function: f(x) = 0 if x < 0, 1 if x >= 0 Note: Cannot be used during training due to zero derivative

ThresholdSymmetric

Symmetric threshold: f(x) = -1 if x < 0, 1 if x >= 0 Note: Cannot be used during training due to zero derivative

Sigmoid

Sigmoid activation function: f(x) = 1 / (1 + exp(-2 * steepness * x)) Output range: (0, 1)

SigmoidSymmetric

Symmetric sigmoid (tanh): f(x) = tanh(steepness * x) Output range: (-1, 1)

Tanh

Hyperbolic tangent: alias for SigmoidSymmetric

Gaussian

Gaussian activation: f(x) = exp(-x * steepness * x * steepness) Output range: (0, 1], peaks at x=0

GaussianSymmetric

Symmetric gaussian: f(x) = exp(-x * steepness * x * steepness) * 2 - 1 Output range: (-1, 1], peaks at x=0

Elliot

Elliott activation: f(x) = ((x * steepness) / 2) / (1 + |x * steepness|) + 0.5 Fast approximation to sigmoid, output range: (0, 1)

ElliotSymmetric

Symmetric Elliott: f(x) = (x * steepness) / (1 + |x * steepness|) Fast approximation to tanh, output range: (-1, 1)

LinearPiece

Bounded linear: f(x) = max(0, min(1, x * steepness)) Output range: [0, 1]

LinearPieceSymmetric

Symmetric bounded linear: f(x) = max(-1, min(1, x * steepness)) Output range: [-1, 1]

ReLU

Rectified Linear Unit (ReLU): f(x) = max(0, x) Output range: [0, โ)

ReLULeaky

Leaky ReLU: f(x) = x if x > 0, 0.01 * x if x <= 0 Output range: (-โ, โ)

Sin

Sine activation: f(x) = sin(x * steepness) / 2 + 0.5 Output range: [0, 1]

Cos

Cosine activation: f(x) = cos(x * steepness) / 2 + 0.5 Output range: [0, 1]

SinSymmetric

Symmetric sine: f(x) = sin(x * steepness) Output range: [-1, 1]

CosSymmetric

Symmetric cosine: f(x) = cos(x * steepness) Output range: [-1, 1]

Implementations

impl ActivationFunction

pub fn name(&self) -> &'static str

Returns the string name of the activation function

pub fn is_trainable(&self) -> bool

Returns whether this activation function can be used during training (i.e., has a computable derivative)

pub fn output_range(&self) -> (&'static str, &'static str)

Returns the output range of the activation function